Make the Scythe AOE harvest ability compatible with EnderCore
