+++ dependencies.gradle
@@ -23,5 +23,6 @@ dependencies {
     // devOnlyNonPublishable("com.github.GTNewHorizons:harvestcraft:1.3.1-GTNH:dev")
     // devOnlyNonPublishable(deobfCurse("pams-harvest-the-nether-231262:2241397"))
     // devOnlyNonPublishable("com.github.GTNewHorizons:Natura:2.8.5:dev")
+    // devOnlyNonPublishable("com.github.GTNewHorizons:EnderCore:0.4.6")
 
 }
